refactor(session): extract resources to own file, fix workspace lookup error handling
Move ResourceStatus, ResourceState, and Resources into resources.go.
Pass workspace directly to NewSessionResponse instead of resolving
name and path separately. Return and handle errors from lookupWorkspace.
