+++ internal/session/session.go
@@ -21,27 +21,6 @@ const (
 	StatusDeleted  SessionStatus = "deleted"
 )
 
-type ResourceStatus string
-
-const (
-	ResourcePending  ResourceStatus = "pending"
-	ResourceCreating ResourceStatus = "creating"
-	ResourceReady    ResourceStatus = "ready"
-	ResourceFailed   ResourceStatus = "failed"
-	ResourceRemoved  ResourceStatus = "removed"
-)
-
-type ResourceState struct {
-	Status ResourceStatus `json:"status"`
-	Error  string         `json:"error,omitempty"`
-}
-
-type Resources struct {
-	Branch   *ResourceState `json:"branch,omitempty"`
-	Worktree *ResourceState `json:"worktree,omitempty"`
-	Tmux     *ResourceState `json:"tmux,omitempty"`
-}
-
 type Session struct {
 	ID              string        `json:"id"`
 	TmuxSessionName string        `json:"tmux_session_name,omitempty"`
@@ -56,30 +35,6 @@ type Session struct {
 	IsAttached      bool          `json:"is_attached"`
 	WorkspaceName   string        `json:"workspace_name,omitempty"`
 	LastUsedAt      time.Time     `json:"last_used_at"`
-}
-
-func (r *Resources) AllReady() bool {
-	if r == nil {
-		return true
-	}
-	for _, rs := range []*ResourceState{r.Branch, r.Worktree, r.Tmux} {
-		if rs != nil && rs.Status != ResourceReady {
-			return false
-		}
-	}
-	return true
-}
-
-func (r *Resources) FirstError() string {
-	if r == nil {
-		return ""
-	}
-	for _, rs := range []*ResourceState{r.Branch, r.Worktree, r.Tmux} {
-		if rs != nil && rs.Error != "" {
-			return rs.Error
-		}
-	}
-	return ""
 }
 
 func SanitizeID(name string) string {
